pyproject.toml
[project]
name = "fast-api-core"
version = "0.1.0"
dependencies = [
    "aerich>=0.9.1",
    "bcrypt>=4.3.0",
    "fastapi>=0.116.1",
    "fastapi-limiter>=0.1.6",
    "passlib>=1.7.4",
    "pydantic-settings>=2.10.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "redis>=6.2.0",
    "tortoise-orm>=0.25.1",
    "uvicorn>=0.35.0",
    "taskipy>=1.14.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "pydantic>=2.11.7",
    "fastapi-admin>=1.0.4",
    "setuptools>=68.0.0",
    "python-dotenv>=1.0.1",
    "python-decouple>=3.8",
    "fastadmin[fastapi,tortoise-orm]>=0.2.22",
]

[tool.aerich]
tortoise_orm = "core.main.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[tool.taskipy.tasks]
test = "pytest"
start = "uvicorn core.main:app --reload"
migrate = "uv run aerich migrate"
upgrade = "uv run aerich upgrade"
init-db = "uv run aerich init-db"
init-aerich = "uv run aerich init -t core.main.TORTOISE_ORM"

[tool.setuptools.packages.find]
where = ["."]
include = ["core*"]
exclude = ["migrations*"]
